MNT: rename modules, idfx setup is now idfx conf
