+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- MNT: rename modules, idfx setup is now idfx conf
+  https://github.com/neutrinoceros/idefix_cli/pull/82
+
 ## [0.5.0] - 2021-23-09
 
 ENH: implement a plugin architecture for commands
@@ -40,7 +45,7 @@ UX: add a more specific error message when IDEFIX_DIR exists but isn't valid.
 - refactor: use temporary ini files in `idfx run` to avoid polluting the work
    directory. Remove "temporary" `*.ini_` filename pattern from `idfx clean`'s
    rules (backwards incompat)
-- refactor: remove error wrapping in `idfx setup` to allow debugging via breakpoints
+- refactor: remove error wrapping in `idfx conf` to allow debugging via breakpoints
 - UX: add default values for the `directory` positional argument in `idfx clean` and `idfx run`
 - UX: display a clean error message in case idefix itself errors out in `idfx run`
