+++ idefix_cli/_commands/clean.py
