+++ idefix_cli/_commands/read.py
