+++ idefix_cli/_commands/stamp.py
