+++ idefix_cli/_commands/write.py
