+++ setup.cfg
@@ -29,7 +29,7 @@ exclude =
 
 [options.entry_points]
 console_scripts =
-    idfx = idefix_cli.main:main
+    idfx = idefix_cli._main:main
 
 [options.extras_require]
 dev =
